% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{bar_cols}
\alias{bar_cols}
\title{Barplot Color Vector}
\format{
An object of class \code{character} of length 6.
}
\usage{
data(bar_cols)
print(bar_cols)
}
\description{
A six-color vector for barplots, containing the hex codes: c("black", "#41afaa", "#466eb4", "#00a0e1", "#e6a532", "#d7642c").
}
\keyword{datasets}
