% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Con_Simul}
\alias{Con_Simul}
\title{Simulate Contingency Table}
\usage{
Con_Simul(
  probs = "equal",
  total_count = 750,
  n_lesion = 3,
  n_Trt. = 5,
  margin_fixed_Trt. = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{probs}{Matrix of probability values created using \code{matrix()}. Each row in the matrix should represent a treatment group and each column a lesion category. All probability values in the matrix should sum to 1. Default = equal probability across all cells.}

\item{total_count}{Total number of counts in the contingency table. Defaults to 750.}

\item{n_lesion}{Number of lesion categories. Ignored if \code{probs} specified. Defaults to 3.}

\item{n_Trt.}{Number of treatment groups. Ignored if \code{probs} specified. Defaults to 5.}

\item{margin_fixed_Trt.}{Whether margins are fixed per treatment group (i.e. fixed number of fish per treatment). Default = FALSE. See \bold{Details} for further information on marginals.}

\item{verbose}{Whether to print the parameters and probability matrix used. Default = TRUE.}
}
\value{
Returns a list containing:\tabular{lll}{
 \code{sim_tab} \tab \tab The simulated contingency table containing counts across different treatment groups (rows) and lesion categories (columns) \cr
 \code{params} \tab \tab The simulation parameters as a vector \cr
 \code{probs} \tab \tab The probability matrix used for simulation \cr
}
}
\description{
Simulate a contingency table with fish counts distributed across \emph{n} lesion categories and \emph{n} treatment groups. Probability values for generating counts in each cell (i.e. each factor level combination) can be assigned using the \code{probs} argument. This function is designed for use in power and/or false positive rate calculations; for details, see \code{Con_Simul_PR()}.
}
\details{
Counts are simulated from a multinomial distribution using \code{rmultinom()}. Counts may be assumed to have a fixed total in the marginals (e.g. per treatment group) or no fixed total in row or column marginals.

For further discussion into the types of marginals in contingency tables, refer to: \url{https://www.uvm.edu/~statdhtx/StatPages/More_Stuff/Chi-square/Contingency-Tables.pdf} and the comments on \bold{Arguments}.
}
\examples{
Con_Tab = Con_Simul(total_count = 750, n_lesion = 3, n_Trt. = 5)
Con_Tab = Con_Simul(probs = matrix(nrow = 2, ncol = 3, c(1/6, 3/12, 1/6, 1/6, 1/6, 1/12)))
}
