% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Con_Simul_PR}
\alias{Con_Simul_PR}
\title{Calculate Positive Rates for Contingency Table}
\usage{
Con_Simul_PR(
  Con_Simul_Object = Con_Simul(),
  add_fisher_exact = FALSE,
  add_ord = FALSE,
  sample_sizes = NULL,
  n_sim = 1000,
  FPR = TRUE
)
}
\arguments{
\item{Con_Simul_Object}{Output from \code{Con_Simul()} with the argument \code{verbose} set to TRUE.}

\item{add_fisher_exact}{Whether to compute positive rates for Fisher's Exact test. May add >1 min of calculation time. Defaults to FALSE.}

\item{add_ord}{Whether to compute positive rates for Wald test on a fitted ordinal regression model. May add >1 min of calculation time. Defaults to FALSE.}

\item{sample_sizes}{A vector of sample sizes over which false positive rates are to be calculated. A sample size is defined as the total number of counts in a contingency table. Defaults to total count received by \code{Con_Simul_Object}.}

\item{n_sim}{Number of contingency tables simulated for each positive rate calculation. Defaults to 1000.}

\item{FPR}{Whether to calculate false positive rate in addition to power. Defaults to TRUE.}
}
\value{
Returns a list containing:\tabular{lll}{
 \code{pos_rate} \tab \tab A dataframe containing positive rates of various tests at specified sample sizes \cr
 \code{eff_mat} \tab \tab The probability matrix used to calculate power  \cr
 \code{null_mat} \tab \tab The probability matrix used to calculate false positive rate \cr
 \code{plot} \tab \tab A plot of positive rates of various tests at specified sample sizes \cr
}
}
\description{
Computes statistical power and optionally false positive rates for tests applied to contingency tables based on Monte Carlo simulations. Specify the simulation process using \code{Con_Simul()}, which serves as input. Positive rates are computed for the Chi-square test and optionally for Fisher's exact test and the Wald test applied to an ordinal regression model.
}
\details{
Power is defined as the percentage of tests yielding positive results (p-value < 0.05) on a set of contingency tables simulated based on the data-generating process and probability matrix specified in \code{Con_Simul()}. The specified probability matrix should represent the parameters of the population where there is a desire to detect a significant effect in the sample. The simulated contingency tables then reflect the sample outcomes of the specified population parameter.

False Positive Rate is defined as the percentage of tests yielding positive results (p-value < 0.05) on contingency tables simulated based on a probability matrix without treatment effect. The odd ratios between lesion scores for all treatment groups are assumed to be the same as that for control (first row in the probability matrix).

P-values for the Chi-square test are computed using \code{stats::chisq.test()} with default parameters. P-values for Fisher's exact test are computed using \code{stats::fisher.test()} with \code{simulate.p.value} set to TRUE, alongside default parameters. P-values for ordinal regression are computed from \code{stats::anova()} applied to the output of \code{ordinal::clm()}.
}
\examples{
## Below I show how we can perform a simple power calculation using this tool.
## Suppose I want to calculate power for Treatment B which halves the lesions in category 2 and 3.
## I then specify the following probability matrix and feed it into Con_Simul():
probs_mat = matrix(nrow = 2, ncol = 3, data = c(1/6, 1/3, 1/6, 1/12, 1/6, 1/12))
sim_tab = Con_Simul(probs_mat)

## Next, I feed the output into Con_Simul_PR():
Con_Simul_PR(sim_tab, sample_sizes = c(50, 100, 150))
## Results: Power is ~55, 86, and 97\% for the Chi-square test using total counts of 50, 100, and 150, respectively.

## The same power for Chi-square test can be calculated using Cohen's omega (w) method which is faster but has its own limitations;
## e.g. assumes one data generating process for the contingency table (the no fixed marginals).
library(pwr)
pwr.chisq.test(w = ES.w2(probs_mat), df = 2, sig.level = 0.05, N = 100)
## Results: Power is 85.6\% for the Chi-square test at the total count of 100.
}
