% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{est_fish_sgr_sd}
\alias{est_fish_sgr_sd}
\title{Estimate Inter-fish SGR SD}
\usage{
est_fish_sgr_sd(
  Wi,
  Wf,
  t,
  sd_limits = c(0, 0.3),
  sd_increments = 0.002,
  n_sim = 1000,
  show_plot = TRUE,
  plot_title = ""
)
}
\arguments{
\item{Wi}{placeholder}
}
\value{
placeholder
}
\description{
placeholder
}
\examples{
#placeholder
}
