% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{est_tank_sgr_sd}
\alias{est_tank_sgr_sd}
\title{Estimate Inter-Tank SGR SD}
\usage{
est_tank_sgr_sd(
  fish_sd,
  fish_num,
  tank_num,
  observed_tank_sd,
  sd_limits = c(0, 0.1),
  sd_increments = 0.001,
  n_sim = 1000,
  plot_title = "",
  show_plot = TRUE
)
}
\arguments{
\item{fish_sd}{placeholder}
}
\value{
placeholder
}
\description{
placeholder
}
\examples{
#placeholder
}
