% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{fpr_check}
\alias{fpr_check}
\title{Check False Positive Rate}
\usage{
fpr_check(
  formula,
  data,
  model_or_test,
  n_sim = 1000,
  return_properties = FALSE
)
}
\arguments{
\item{formula}{A one-factor formula with the dependent variable (y) on the left hand side, a tiddle (~), and then the independent variable (x), e.g. y ~ x.}

\item{data}{Relevant dataframe representing the sampling design (sample sizes in different x levels).}

\item{model_or_test}{A string specifying the model or test to check false positive rate for. Currently supports: "logistic", "chisq", "anova" (or the equivalent "lm").}

\item{n_sim}{Numeric describing the number of p-values to simulate.}

\item{return_properties}{TRUE/FALSE describing whether to return characteristics for each simulation. Intended use is for troubleshooting.}
}
\value{
A printout describing the false positive rate. A histogram of the relevant test-statistic and p-value under the null. If \code{return_properties == TRUE}, a list containing the vector of p-values, vector of test-statistics, and model object or others deemed important.
}
\description{
Assesses the False Positive Rate or fraction of p values below 0.05 by simulating p-values given that the null is true (no effect of "x" (usually treatment) specified in \code{formula}). Uses the sampling design (sample sizes per x-level) indicated by the supplied \code{data}. Simulation involves resampling of the vector of values in "y" with replacement. Outputs a histogram showing the null distribution of p-values and test-statistic.
}
\examples{
#Check to see if false positive rate is acceptable (0.05) for ANOVA on iris dataset:
data(iris)
fpr_check(Sepal.Length ~ Species, iris, "anova", n_sim = 2000)

#Now check to see fpr for a low count dataset:
test_df = data.frame(Trt.ID = rep(c("A", "B"), times = 20),
                     Pathology = c(rep(1, 34), rep(0, 6)))
table(test_df)

fpr_check(Pathology ~ Trt.ID, test_df, "chisq") |>
  suppressWarnings() #Warning: Chisquare approximation may be incorrect

fpr_check(Pathology ~ Trt.ID, test_df, "logistic")
}
