% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{haz_db_ex}
\alias{haz_db_ex}
\title{Example Hazard Data}
\format{
A data frame containing 54 rows and 3 columns:\tabular{lll}{
 \code{Trt.ID} \tab \tab A label for the treatment group used in creating this reference hazard dataframe \cr
 \code{Hazard} \tab \tab Hazard values (rates) \cr
 \code{Time} \tab \tab Time / TTE in days post challenge. \cr
}
}
\usage{
data(haz_db_ex)
View(haz_db_ex)
}
\description{
A reference hazard dataframe created using \code{Surv_Plots(..., dailybin = TRUE, data_out = TRUE)$Hazard_DB} which uses \code{bshazard::bshazard()} inside. Contains hazard rates over time.
}
\keyword{datasets}
