% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{knife}
\alias{knife}
\title{Truncate Number}
\usage{
knife(x, digits = 4)
}
\arguments{
\item{x}{Vector of numeric values}

\item{digits}{Number below which value is converted to scientific notation.}
}
\value{
Truncated values
}
\description{
Truncate a decimal by rounding or converting to scientific notation. Originally intended for p-values.
}
