% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Label_Gen}
\alias{Label_Gen}
\title{Generate Texts for Labels}
\usage{
Label_Gen(
  input_list,
  sort_by = NULL,
  n_col = 6,
  fill_by_row = TRUE,
  save_name = NULL
)
}
\arguments{
\item{input_list}{A list of named variables, each containing one or more text/number(s). See \bold{Examples} for examples.}

\item{sort_by}{A value or vector representing the variable(s) to sort the output by. For each variable, sorts according to the order of text in the variable. When multiple variables is given, prioritizes sorting based on the order of variables; leftmost = highest priority. Defaults to NULL where sorting is based on \code{input_list} orders.}

\item{n_col}{The number of columns in the output table. It should match the number of columns in the label paper. Defaults to 6.}

\item{fill_by_row}{Whether combinations should fill the output table by row (otherwise column). Defaults to TRUE.}

\item{save_name}{Name of the saved .csv. Defaults to NULL where the file name is "Label_Gen" and today's date (YYYY-MM-DD).}
}
\value{
A .csv containing all possible combinations. Additionally, a printout describing the .csv file name and location. Another printout describing the total number of labels / combinations created.
}
\description{
Combines texts specified in a list. List should contain multiple variables that holds the texts in a value or vector format. All combinations of texts across variables are computed. Output text is sorted in order of variables given in the list (default behavior) or as specified using \code{sort_by} argument. The output combinations are tabulated and saved in a .csv in your working directory.
}
\examples{
# Summarize the input variables in a list
input_variables = list(Time = c("Baseline", "1wpv"),
                       Animal = c("Oysters", "Lobsters"),
                       Tissue = c("Meat", "Shell", "Water", "Head"),
                       Replic_num = 1:3)

# Run Label_Gen() using the input variables.
Label_Gen(input_list = input_variables,
          sort_by = c("Time", "Animal", "Tissue"),
          n_col = 6,
          fill_by_row = TRUE,
          save_name = NULL)
}
\seealso{
\href{https://sean4andrew.github.io/safuncs/reference/Label_Gen.html}{Link} for web documentation.
}
