% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{mort_db_ex}
\alias{mort_db_ex}
\title{Example Mort Data}
\format{
A data frame containing 399 rows and 4 columns:\tabular{lll}{
 \code{Tank.ID} \tab \tab Unique labels for the different tanks in the study \cr
 \code{Trt.ID} \tab \tab Unique labels for the different treatments in the study \cr
 \code{TTE} \tab \tab Time to Event. In this dataset, TTE = days post challenge \cr
 \code{Status} \tab \tab Value indicating what happened at TTE. In this dataset, Status = 1 indicating all events are death \cr
}
}
\usage{
data(mort_db_ex)
View(mort_db_ex)
}
\description{
An example mortality dataframe that can be accepted by \code{Surv_Gen()}.
}
\keyword{datasets}
