% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{multivar_db_ex}
\alias{multivar_db_ex}
\title{Example Multivariate Data}
\format{
An object of class \code{data.frame} with 48 rows and 9 columns.
}
\usage{
data(multivar_db_ex)
View(multivar_db_ex)
}
\description{
An example dataset representing the mucus chemistry of fish exposed to different treatments. Mucus chemistry is described by several variables from columns 2 to 8. In column 9, a fake factor "Fruits" was added for use as an example second factor. In column 4, row 10, the available value was replaced by NA to use as an example missing value.
}
\keyword{datasets}
