% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{nasif_cols}
\alias{nasif_cols}
\title{Nasif's Color Vector}
\format{
An object of class \code{character} of length 7.
}
\usage{
data(nasif_cols)
print(nasif_cols)
}
\description{
A vector containing the hex codes: c("#009B00", "#0000FF", "#FFA900", "#9B009B", "#40E0D0", "#FF0000", "#FFD700").
}
\keyword{datasets}
