% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{onda_cols1}
\alias{onda_cols1}
\title{Primary ONDA Color Vector}
\format{
An object of class \code{character} of length 3.
}
\usage{
data(onda_cols1)
print(onda_cols1)
}
\description{
A vector containing the hex codes for the primary theme of colors for ONDA: c("#004750", "#006B68", "#65BBB0").
}
\keyword{datasets}
