% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{onda_cols2}
\alias{onda_cols2}
\title{Secondary ONDA Color Vector}
\format{
An object of class \code{character} of length 6.
}
\usage{
data(onda_cols1)
print(onda_cols1)
}
\description{
A vector containing the hex codes for the secondary theme of colors for ONDA: c("#163029", "#80C7BC", "#AFE2E3", "#D0D0AA", "#CEDFD7", "#FFFFFF").
}
\keyword{datasets}
