% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{path_db_ex}
\alias{path_db_ex}
\title{Example Pathology Data}
\format{
A data frame containing 591 rows and 12 columns: \tabular{lll}{
 \code{Trt.ID} \tab \tab Unique label representing the treatment group associated with a row of data \cr
 \code{Tank.ID} \tab \tab Unique label representing each different tank \cr
 \code{misc...} \tab \tab Various pathological signs (columns 3-12) \cr
}
}
\usage{
data(path_db_ex)
View(path_db_ex)
}
\description{
A dataframe containing Trt.ID, Tank.ID, and various (13) pathological signs containing binary data (Y/N). Contains "dirty" entries (e.g. N/AP, ME).
}
\keyword{datasets}
