% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Path_Summary}
\alias{Path_Summary}
\title{Create Summary Statistics from Pathology Data}
\usage{
Path_Summary(
  path_db,
  path_cols,
  factors = c("Trt.ID"),
  p_adj = "BH",
  contrast_out = TRUE
)
}
\arguments{
\item{path_db}{A cleaned pathology dataframe such as that from \code{Path_Gen()}. Must contain at least one column as the factor, and one column for a pathological sign. Multiple signs  path}

\item{path_cols}{A numeric vector representing the column indices of pathological signs.}

\item{factors}{A character vector specifying the factor(s) relevant in the dataframe. Defaults to "Trt.ID", Only the first factor is considered in statistical tests, however, summary statistics such as prevalence, standard errors, and n, are calculated for every possible combination of factor level.}

\item{p_adj}{A string specifying the p-value adjustment method to account for multiple comparisons. Defaults to "BH". Method based on \code{p.adjust()}.}

\item{contrast_out}{A logical (TRUE/FALSE) specifying whether to return an additional dataframe containing the p-value based on pairwise comparison of groups/level of the first specified factor in \code{factors}. Defaults to TRUE.}
}
\value{
A dataframe or list (when \code{contrast_out = TRUE}) containing summary statistics and results from univariate hypothesis test.
}
\description{
Takes a cleaned pathology dataframe and summarizes, for each pathological sign, its prevalence, standard error, and n across each unique combination of factors. P-value and letter groups from logistic regression and a wald-type test with p-value adjustments are also produced; the logistic regression model and letters considers only the first specified factor and ignores all others.
}
\examples{
#placeholder
}
