% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Path_Table}
\alias{Path_Table}
\title{Create Flextables from Prevalence Data}
\usage{
Path_Table(path_sum, digits = 0, prev_append = c("se_plus-minus"))
}
\arguments{
\item{path_sum}{A list or dataframe from \code{Path_Summary()}.}

\item{digits}{A number specifying the decimal places for the prevalence values in the output flextable. Defaults to 0.}

\item{prev_append}{A string specifying the suffix to append to prevalence values in the output table. Suffix may be the standard error or counts of presence. Suffix may be presented inside brackets or after a +/- sign. Choose one of the following arguments: "se_brackets", "se_plus-minus", "presence_brackets". Defaults to "se_plus-minus".}
}
\value{
A list that contains, at minimum, 2 flextable objects of the summary statistics (from \code{Path_Summary()}) in wide and long format. If a contrast dataframe from \code{Path_Summary()} is also included in the \code{path_sum} argument, then the output list contains additionally a flextable for the contrast dataframe.
}
\description{
Converts the dataframe(s) from \code{Path_Summary()} into flextable objects in wide and long formats. Flextable can then be printed to word using \code{officer} package.
}
\examples{
#Generate summarized pathology dataframe:
data(path_db_ex)
path_db = Path_Gen(path_db = path_db_ex,
                   rel_cols = 1,
                   path_cols = 3:12)

path_sum = Path_Summary(path_db = path_db,
                        factor = c("Trt.ID"),
                        path_cols = 2:11,
                        contrast_out = TRUE)

path_tables = Path_Table(path_sum = path_sum,
                         digits = 1, #1 decimal places for prevalence values
                         prev_append = c("se_brackets")) #append standard errors (in brackets) to prevalence values

#View pathology table (wide format):
path_tables$wide

#View pathology table (long format):
path_tables$long

#View pathology contrast table:
path_tables$contrast

#Print results in word docx:
library(officer)
library(flextable)
results_doc = read_docx() |>
  body_add_flextable(value = path_tables$wide, align = "left", topcaption = TRUE) |>
  body_add_break() |>
  body_add_flextable(value = path_tables$long, align = "left", topcaption = TRUE) |>
  body_add_break() |>
  body_add_flextable(value = path_tables$contrast, align = "left", topcaption = TRUE)
print(results_doc, target = paste0("ONDA Pathology Analyses ", format(Sys.Date(), "\%d\%b\%Y"),".docx"))
}
