% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Pow_Simul_Mult}
\alias{Pow_Simul_Mult}
\title{Positive Rates for Contingency Tables}
\usage{
Pow_Simul_Mult(
  Simul_Mult_Object = Simul_Mult(),
  add_fisher_exact = FALSE,
  add_ord = FALSE,
  sample_sizes = NULL,
  n_sim = 1000,
  FPR = TRUE
)
}
\arguments{
\item{Simul_Mult_Object}{Output from \code{Simul_Mult()} with the argument \code{verbose} set to TRUE.}

\item{add_fisher_exact}{Whether to additionally calculate the positive rates for Fisher's Exact test. May add >1 min of calculation time. Defaults to FALSE.}

\item{add_ord}{Whether to additionally calculate the positive rates for Wald test on a fitted ordinal regression model. May add >1 min of calculation time. Defaults to FALSE.}

\item{sample_sizes}{A vector of sample sizes over which false positive rates are to be calculated. A sample size is defined as the total number of counts in a contingency table. Defaults to the total count passed onto \code{Simul_Mult_Object}.}

\item{n_sim}{Number of simulated datasets for every calculated positive rate. Defaults to 1000.}

\item{FPR}{Whether to calculate false positive rate in addition to power. Defaults to FALSE.}
}
\value{
placeholder
}
\description{
Calculates power and optionally the false positive rates of statistical tests on contingency tables. The data generation process for the contingency table is specified using \code{Simul_Mult()} which is taken as input. Positive rates are calculated for the Chi-square test and optionally for the Fisher's exact and the Wald test on an ordinal regression model.
}
\details{
Test
}
\examples{
placeholder
}
