% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{sc_db_ex}
\alias{sc_db_ex}
\title{Example Lesion Scores Data}
\format{
A data frame containing 1012 rows (each representing one fish) and 9 columns: \tabular{lll}{
 \code{Timepoint} \tab \tab Unique label representing the timepoint associated with a row of data (i.e. one fish) \cr
 \code{Trt.ID} \tab \tab Unique label representing the treatment group associated with a row of data \cr
 \code{Tank.ID} \tab \tab Unique label representing each different tank \cr
 \code{Mouth} \tab \tab Category (must be 0-4) of severity for the fish based on its mouth lesion(s) \cr
 \code{Gill} \tab \tab Category (must be 0-2) of severity for the fish based on its gill lesion(s) \cr
 \code{Skin1} \tab \tab Count of skin lesions Category 1 \cr
 \code{Skin2} \tab \tab Count of skin lesions Category 2 \cr
 \code{Skin3} \tab \tab Count of skin lesions Category 3 \cr
 \code{Skin3R} \tab \tab Count of skin lesions Category 3R \cr
}
}
\usage{
data(sc_db_ex)
View(sc_db_ex)
}
\description{
A dataframe of lesion scores that can be accepted by the template script for scores analyses. Except for \emph{Tank.ID}, the columns described below are necessary.
}
\keyword{datasets}
