% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{silencer}
\alias{silencer}
\title{Silence Output}
\usage{
silencer(x)
}
\arguments{
\item{x}{}
}
\value{
Object of interest
}
\description{
Hide output from R console by applying \code{sink(tempfile())}
}
\examples{
Placeholder

}
