% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Simul_Mult}
\alias{Simul_Mult}
\title{Simulate a Contingency Table}
\usage{
Simul_Mult(
  probs = "equal",
  total_count = 750,
  n_lesion = 3,
  n_Trt. = 5,
  margin_fixed_Trt. = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{probs}{Matrix of probability values created using \code{matrix()}. Each row in the matrix should represent a treatment group and each column a lesion category. All probability values in the matrix should sum to 1. Default = equal probability across all cells.}

\item{total_count}{Total number of counts in the contingency table. Defaults to 750.}

\item{n_lesion}{Number of lesion categories. Ignored if \code{probs} specified. Defaults to 3.}

\item{n_Trt.}{Number of treatment groups. Ignored if \code{probs} specified. Defaults to 5.}

\item{margin_fixed_Trt.}{Whether margins are fixed per treatment group (i.e. fixed number of fish per treatment). Default = FALSE. See \bold{Details} for further information on marginals.}

\item{verbose}{Whether to print the parameters and probability matrix used. Default = TRUE.}
}
\value{
A list consisting of the simulated contingency table with counts in each cell for the different treatment groups (as rows) and lesion categories (as columns), the table parameters, and the probability matrix for the data generation process.
}
\description{
Simulate a contingency table with fish counts distributed across \emph{n} lesion categories and \emph{n} treatment groups. Probability values for generating counts in each cell (i.e. each factor level combination) can be assigned using the \code{probs} argument. This function is designed for use in power and/or false positive rate calculations; for details, refer to \code{Pow_Simul_Mult()}.
}
\details{
Counts are simulated from a multinomial distribution using \code{rmultinom()}. Counts may be assumed to have a fixed total in the marginals (e.g. per treatment group) or no fixed total in row or column marginals.

For further discussion into the types of marginals in contingency tables, refer to: \url{https://www.uvm.edu/~statdhtx/StatPages/More_Stuff/Chi-square/Contingency-Tables.pdf} and the comments on \bold{Arguments}.
}
\examples{
Con_Tab = Simul_Mult(total_count = 750, n_lesion = 3, n_Trt. = 5)
Con_Tab = Simul_Mult(probs = matrix(nrow = 2, ncol = 3, c(1/6, 3/12, 1/6, 1/6, 1/6, 1/12)))
}
