% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{starting_fish_count_ex}
\alias{starting_fish_count_ex}
\title{Example Starting Count Data}
\format{
A data frame containing 4 treatments, 12 tanks, and starting fish counts for each tank laid out in three columns named "Trt.ID", "Tank.ID", and "starting_fish_count"
}
\usage{
data(starting_fish_count_ex)
View(starting_fish_count_ex)
}
\description{
A dataframe of starting fish counts that can be accepted by \code{Surv_Gen()}.
}
\keyword{datasets}
