% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{surv_db_ex}
\alias{surv_db_ex}
\title{Example Survival Data}
\format{
A data frame containing 1200 rows and 4 columns:\tabular{lll}{
 \code{Tank.ID} \tab \tab Unique labels for the different tanks in the study \cr
 \code{Trt.ID} \tab \tab Unique labels for the different treatments in the study \cr
 \code{TTE} \tab \tab Time to Event. In this dataset, TTE = days post challenge \cr
 \code{Status} \tab \tab Value indicating what happened at TTE. In this dataset, Status = 1 or 0 indicating death or survival, respectively \cr
}
}
\usage{
data(surv_db_ex)
view(surv_db_ex)
}
\description{
A complete survival dataset with survivors, based of \code{mort_db_ex}. Ready for proper survival analysis.
}
\keyword{datasets}
