% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Gen}
\alias{Surv_Gen}
\title{Generate Survivor Data}
\usage{
Surv_Gen(
  mort_db,
  today_tte,
  starting_fish_count,
  tank_without_mort = NULL,
  trt_without_mort = NULL
)
}
\arguments{
\item{mort_db}{A mort dataframe as described in \bold{Details}.}

\item{today_tte}{The time to event assigned to every generated row of survivor data.}

\item{starting_fish_count}{Value representing the starting number of fish per tank.}

\item{tank_without_mort}{Tank IDs not in the mort dataframe but needing survivor data generation. Input a vector of strings.}

\item{trt_without_mort}{Treatment IDs corresponding to \code{tank_without_mort} in the same order. Input a vector of strings.}
}
\value{
A dataframe produced by combining the input mort data and generated rows of survivor data.
}
\description{
Returns a completed survival dataset that include rows for every survivor given the starting number of fish and mortality count. Mortality per tank is supplied through the argument \code{mort_db}. To generate survivor data for tanks absent from \code{mort_db} (e.g. those without mortalities), specify \code{tank_without_mort} and \code{trt_without_mort} arguments.
}
\details{
The mort dataframe input should consist of one row of data for every dead/removed fish and the following 4 columns at minimum:
\itemize{
\item "Trt.ID" = Labels for treatment groups in the study.
\item "Tank.ID" = Labels for tanks in the study (each tank must have a unique label).
\item "TTE" = Time to Event. Event could be fish death or being sampled and removed depending on "Status".
\item "Status" = Value indicating what happened at TTE. 1 for dead fish, 0 for those sampled and removed.
}

For an example dataset, view \code{data(mort_db_ex)}.
}
\examples{
Surv_Gen(mort_db = mort_db_ex,
         today_tte = 54,
         starting_fish_count = 100,
         tank_without_mort = c("C99", "C100"),
         trt_without_mort = c("A", "B"))

}
