% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\docType{data}
\name{surv_sim_db_ex}
\alias{surv_sim_db_ex}
\title{Example Simulated Survival Object}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(surv_sim_db_ex)
View(surv_sim_db_ex$surv_simul_db)
View(surv_sim_db_ex$surv_pop_db)
View(surv_sim_db_ex$surv_plots)
}
\description{
A list generated from \code{Surv_Simul()} containing the following elements: a simulated survival dataset (\code{surv_simul_db}), the dataset describing the survival characteristics of the population (\code{surv_pop_db}) and a plot illustrating both (\code{surv_plots}). The specified simulation parameters (arguments) in \code{Surv_Simul()} are \code{fish_num_per_tank = 100}, \code{tank_num_per_trt = 4}, \code{treatments_hr = c(1, 0.7, 0.5)}, \code{logHR_sd_intertank = 0}, \code{n_sim = 10}, \code{plot_out = TRUE}.
}
\keyword{datasets}
