% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Simul}
\alias{Surv_Simul}
\title{Simulate Survival Data}
\usage{
Surv_Simul(
  Haz_DB,
  fish_num_per_tank = 100,
  tank_num_per_trt = 3,
  Treatments_HR = c(1, 1, 1, 1),
  logHR_sd_intertank = 0,
  integer_days = TRUE
)
}
\arguments{
\item{Haz_DB}{Description not yet available}

\item{fish_num_per_tank}{Description not yet available}

\item{tank_num_per_trt}{Description not yet available}

\item{Treatments_HR}{Description not yet available}

\item{logHR_sd_intertank}{Description not yet available}

\item{integer_days}{Description not yet available}
}
\value{
A simulated survival dataset.
}
\description{
Not yet available
}
\examples{
Not yet available
}
