% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{xlsx_row2coln}
\alias{xlsx_row2coln}
\title{Set First Row as Column Headers}
\usage{
xlsx_row2coln(x)
}
\arguments{
\item{x}{A tibble / dataframe object. Initially designed for the output of \code{readxl::read_xlsx()}.}
}
\value{
A dataframe where the previous first rows are now column names
}
\description{
Transform first row of the data into column names and subsequently remove that row.
}
