% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{GG_Colour_Hue}
\alias{GG_Colour_Hue}
\title{Get Default Colours by ggplot}
\usage{
GG_Colour_Hue(n)
}
\arguments{
\item{n}{Number of colour groups}
}
\value{
Returns a vector representing the default colour codes assigned to each group by ggplot.
}
\description{
Not my function but it is useful so here it is! \href{Origin}{https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
}
\examples{

# Get colour codes used for 6 categorical groups
GG_Colour_Hue(6)

}
