% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{MultiVar}
\alias{MultiVar}
\title{Analyze and Visualize Multivariate Data}
\usage{
MultiVar(
  multivar_db,
  values_cols,
  factors_cols,
  factors_pool = c("col1", "col2"),
  factors_facet = "none",
  pca_ellipse = c("confidence"),
  pca_labels = NULL,
  missing_method = "imputation",
  scale = TRUE,
  center = TRUE,
  boxplot_filled = TRUE,
  boxplot_x_angle = NULL,
  boxplot_x_wrap = NULL,
  boxplot_x_lab = FALSE,
  boxplot_legend = TRUE,
  colours = NULL,
  univariate_tests = TRUE,
  plot_out_png = FALSE,
  plot_out_pptx = FALSE,
  plot_out_R = FALSE
)
}
\arguments{
\item{multivar_db}{A dataframe with two types of columns. The first holds numeric values of the multivariate outcome variables, with each column containing one variable. The second type holds categories (factor levels), with each column containing one factor. An example dataframe can be viewed by running \code{View(data(multivar_db_ex))} in the R console.}

\item{values_cols}{A numeric vector specifying the order number of columns containing the outcome variables.}

\item{factors_cols}{A numeric vector specifying the order number of columns containing the factors. Maximum of two numbers (i.e. factors).}

\item{factors_pool}{A character vector indicating the factors which levels are to be pooled across in additional plots. Choose any combination of "col1" and "col2" which refers to the first and second column in \code{factors_cols}. Defaults to c("col1", "col2").}

\item{factors_facet}{A character vector indicating the factors which levels are to be faceted across in additional plots. Choose any combination of "col1" and "col2" which refers to the first and second column in \code{factors_cols}. Defaults to "none" which creates no additional plots.}

\item{pca_ellipse}{A character vector representing the type of ellipses to draw in PCA plots. Generates a plot for every specified type. Choose any combination of "confidence", "distribution", "convexhull", and/or "none". "confidence" draws ellipses representing the 95 percent confidence interval about the center of multivariate normal data (principal component scores); drawn using \code{ggpubr::stat_conf_ellipse()}. "distribution" represents ellipses expected to cover 95 percent of all multivariate normal data; drawn using \code{ggplot2::stat_ellipse()} with argument \code{type = "norm"}. "convexhull" represents the smallest convex polygon enclosing all points; drawn using \code{ggpubr::stat_chull()}. For plots without ellipses, include "none". Defaults to c("confidence").}

\item{pca_labels}{A character vector representing the labels to draw in PCA plots. Choose any combination of "ind" and/or "var". "ind" represents individual point labels by their row number. "var" represents variable loadings drawn as arrows; the arrow length and direction are calculated as in \code{factoextra::fviz_pca_biplot()}. Defaults to NULL (no labels drawn).}

\item{missing_method}{A string representing the method to address missing values in \code{values_cols}. Choose from "imputation" or "na_omit". "imputation" fills in missing values with values created (imputed) based on the correlation between variables essentially; accomplished using \code{missMDA::imputePCA()} with the ncp parameter \code{missMDA::estim_ncpPCA()}. "na_omit" removes entire rows of data when at least one NA value is present. This method may result in a significant loss of data. Defaults to "imputation". The choice of \code{missing_method} would affect PCA, LDA, and MANOVA results but likely only to a small degree with few missing values. Has no impact on boxplots and ANOVAs.}

\item{scale}{TRUE/FALSE indicating whether to scale variable values (such that SD = 1 for each variable) before PCA or LDA. A common procedure in the z-score normalization of values that commonly precede PCA. It is not recommended to set this to FALSE, unless justified. Defaults to TRUE.}

\item{center}{TRUE/FALSE indicating whether to center variable values (such that mean = 0 for each variable) before PCA or LDA. A common procedure in the z-score normalization of values that commonly precede PCA. It is not recommended to set this to FALSE, unless justified. Defaults to TRUE.}

\item{boxplot_filled}{TRUE/FALSE indicating whether to color the insides of boxplots and points (i.e. fill them). If FALSE, boxplots and points are hollow with colored borders using \code{colour} in \code{ggplot2::aes()}, instead of using \code{fill}. Defaults to TRUE.}

\item{boxplot_x_angle}{A number describing the degree of tilt in the x-axis labels of the boxplots. Defaults to NULL (horizontal labels).}

\item{boxplot_x_wrap}{The maximum number of characters on a single line that would be split if a space bar is available between them. Defaults to NULL (no text wrapping).}

\item{boxplot_x_lab}{TRUE/FALSE indicating whether to include a title for the x-axis of the boxplots. Defaults to FALSE.}

\item{boxplot_legend}{TRUE/FALSE indicating whether to incude the legend for the different groups in the boxplots. Defaults to TRUE}

\item{colours}{A named character vector specifying the colors to use for different factor levels. E.g. For a factor with levels "A", "B", and "C", the \code{colours} vector may look like \code{c('A' = "brown", B = 'blue', C = '#f8e723')}. Defaults to NULL (default ggplot2 colours).}

\item{plot_out_png}{TRUE/FALSE indicating whether to save plots as .png in the working directory. Defaults to FALSE.}

\item{plot_out_pptx}{TRUE/FALSE indicating whether to save plots as editable forms in .pptx in the working directory. Defaults to FALSE.}

\item{plot_out_R}{TRUE/FALSE indicating whether to output plots as ggplot2 objects in a list in R. Defaults to FALSE.}
}
\value{
Returns a Word document containing at least the following types of results (in the given order):

\strong{Multivariate Analyses:}

\enumerate{
\item PC plot(s) illustrating the separation (if any) of PC scores between groups
\item Correlation plot between individual variables and principal components
\item Correlation table summarizing pearson correlation coefficients between all pairs of variables and principal components
\item Contribution table summarizing contribution values of variables to principal components
\item Contribution table summarizing contribution values of variables to major linear discriminants (i.e. to the separation between groups)
\item MANOVA table summarizing statistical evidence for any effect of factor(s)
}

\strong{Univariate Analyses:}

\enumerate{
\item Boxplot(s) comparing values of individual variables across groups
\item ANOVA results table (Two-Factor or One-Factor)
}
}
\description{
Expedite multivariate analysis and gain comprehensive insights to the data through PCA, LDA, and MANOVA. Outputs include tables of statistical results, PCA plots, and Boxplots of each dependent variable. Plots can be customized using \code{pca_} and \code{boxplots_} prefixed arguments. Supports one- or two-factor analyses. In two-factor analyses, additional plots may be created with facets and/or pooling of values across levels of a selected factor, chosen using the arguments \code{factors_pool} and \code{factors_facet}. By default, saves results in a Word document, but allows exports as .pptx, .png, and/or R objects for further edits. A tutorial on how to generate various outputs from \code{MultiVar()} is available under \bold{Examples}.
}
\details{
Several functions published on CRAN are used by \code{MultiVar()} for various types of statistical analyses:

\describe{
 \item{PCA}{Uses \code{stats::prcomp()}. Pre-PCA, outcome variables may be scaled or centered using \code{scale} and \code{center} arguments, respectively. Missing values may have been imputed using \code{missMDA::imputePCA()} and ncp parameter \code{ncp = missMDA::estim_ncpPCA()$ncp}.}
 \item{LDA}{Uses \code{MASS::lda()}. Pre-LDA, values may be scaled, centered, and missing values imputed, with the same methods as described for PCA.}
 \item{MANOVA}{Uses \code{stats::manova()} and subsequently \code{car::Anova()} with the argument \code{type = 3} - relevant for analyses with two or more factors. Missing values may have been imputed with the same method described for PCA. By default, \code{stats::manova()} omits all rows with missing values. To achieve this in \code{MultiVar()}, set \code{missing_method} argument to "na_omit".}
 \item{PERMANOVA}{Output available upon request. Uses \code{RVAideMemoire::adonis.II()} with the argument \code{method = "euclidean"}. Missing values may be imputed using the same method described for PCA. By default, \code{RVAideMemoire::adonis.II()} omits all rows with missing values. To achieve this in \code{MultiVar()}, set \code{missing_method} argument to "na_omit".}
 \item{ANOVA}{Uses \code{stats::anova()} and subsequently \code{car::Anova()} with the argument \code{type = 3} - relevant for analyses with two or more factors.}
}
}
\examples{
# Below is a brief tutorial to help you get various outputs from MultiVar()!

# Lets start simple, with the aim of creating a report for a one-factor dataset
# called iris:
data(iris)

# Data structure:
head(iris, n = 5)

# We can see that columns #1 to #4 contain dependent variables, while column #5 holds
# the factor. Simply feed this information into Multivar()!
MultiVar(multivar_db = iris,
         values_cols = 1:4,
         factors_cols = 5)

# A report is now saved in your working directory, check it out! You can locate your
# working directory using:
getwd()

# Suppose you wanted a more equipped PCA plot. We can customize our plots using a
# set of arguments indicated with the proper prefix; 'pca_' for PCA plots and
# 'boxplots_' for Boxplots:
MultiVar(multivar_db = iris,
         values_cols = 1:4,
         factors_cols = 5,
         pca_ellipse = c("none", "confidence", "distribution", "convexhull"),
         pca_labels = c("var"),
         plot_out_R = TRUE)$pca$distribution

# If you want to customize the plot beyond the capabilities of MultiVar(), save them
# as editable forms in .pptx or manipulate them further in R:
Sepal_L_box = MultiVar(multivar_db = iris,
                       values_cols = 1:4,
                       factors_cols = 5,
                       boxplot_x_angle = 45,
                       plot_out_pptx = TRUE,
                       plot_out_R = TRUE)$box$none$'Sepal Length'

Sepal_L_box + ggthemes::theme_tufte() # Use the thufte theme

ggplot2::ggsave(filename = "box_SepalLength.tiff",
                width = 5,
                height = 5,
                dpi = 600) # Save with desired plot dimensions and resolution

# MultiVar() can also handle missing values which are common in real life data. By
# default it imputes (creates) missing values based on the correlational properties
# between dependent variables (essentially). Alternatively, rows with missing values
# can be omitted using missing_method = "na_omit". Below is a demo showing the use
# of the imputation method.
iris[1:25, 1] = NA # replace half of Setosa's Sepal.Length with NA

MultiVar(multivar_db = iris,
         values_cols = 1:4,
         factors_cols = 5,
         missing_method = "imputation",
         pca_ellipse = "distribution",
         plot_out_R = TRUE)$pca$distribution

# The plot is not clearly different than before which indicates the imputed values
# substitute NAs effectively, without clear bias.

## TWO FACTOR CASES
# Real data often have two factors, and MultiVar() produces plots tailored to that.
# At minimum, MultiVar() produces plots with the interaction between two factors as
# x-axis labels or legend labels. This can lead to a lot of levels in the axis or
# legends which look unsightly. Run the code below to see how this looks like! Below
# I am using an example fish mucus dataset with one real factor (Treatment) and one
# fake (Fruits).
data(multivar_db_ex)

head(multivar_db_ex, n = 5)

MultiVar(multivar_db = multivar_db_ex,
         values_cols = 2:8,
         factors_cols = c(1, 9),
         factors_pool = "none",
         factors_facet = "none")

# Check out the report in your working directory!

# Additional plots can be created with facets or pooling of values across levels of a
# factor. Below we explore the default parameters which consists of 'factors_pool =
# c('col1', 'col2')' and its output:
MultiVar(multivar_db = multivar_db_ex,
         values_cols = 2:8,
         factors_cols = c(1, 9),
         factors_pool = c("col1", "col2"),
         plot_out_R = TRUE)$pca$confidence[c("pooled1", "pooled2")]

# The "col1" specification instructs MultiVar() to create a plot with pooling across
# levels of the first factor specified in 'values_cols', while "col2" refers to the
# second factor.

# Plots with facets can also be created. Below, we facet against levels of the second
# factor (Fruits) which is considered as a nuisance factor in this case:
MultiVar(multivar_db = multivar_db_ex,
         values_cols = 2:8,
         factors_cols = c(1, 9),
         factors_pool = c("col1", "col2"),
         factors_facet = c("col2"))$pca$confidence["facet2"]

# Check your word document output and you will see that it contains facets and pooling
# for both PCA and boxplots. End of tutorial. Hope it helps!
}
\seealso{
\href{https://sean4andrew.github.io/safuncs/reference/MultiVar.html}{Link} for web documentation.
}
