% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Path_Table}
\alias{Path_Table}
\title{Create Flextables from Prevalence Data}
\usage{
Path_Table(path_sum, digits = 0, prev_append = c("se_plus-minus"))
}
\arguments{
\item{path_sum}{A list or dataframe from \code{Path_Summary()}.}

\item{digits}{A number specifying the decimal places for the prevalence values in the output flextable.}

\item{prev_append}{A string specifying the suffix to append to prevalence values in the output table. Suffix may be the standard error or counts of presence. Suffix may be presented inside brackets or after a +/- sign. Choose one of the following arguments: "se_brackets", "se_plus-minus", "presence_brackets". Defaults to "se_plus-minus".}
}
\value{
A list that contains, at minimum, 2 flextable objects of the summary statistics (from \code{Path_Summary()}) in wide and long format. If a contrast dataframe from \code{Path_Summary()} is also included in the \code{path_sum} argument, then the output list contains additionally a flextable for the contrast dataframe.
}
\description{
Converts the dataframe(s) from \code{Path_Summary()} into flextable objects in wide and long formats.
}
\examples{
#placeholder
}
