% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Pow_Simul_Mult}
\alias{Pow_Simul_Mult}
\title{Positive Rates for Contingency Tables}
\usage{
Pow_Simul_Mult(
  Simul_Mult_Object = Simul_Mult(),
  n_sim = 1000,
  vec_total_count = NULL,
  add_ord = FALSE,
  add_fisher_exact = FALSE,
  FPR = TRUE
)
}
\arguments{
\item{Simul_Mult_Object}{placeholder}

\item{n_sim}{placeholder}

\item{vec_total_count}{placeholder}

\item{add_ord}{placeholder}

\item{add_fisher_exact}{placeholder}

\item{FPR}{placeholder}
}
\value{
placeholder
}
\description{
Calculates power and/or false positive rates for statistical tests on contingency tables generated using a user-defined process. Specify the data generating process using \code{Simul_Mult()} which is taken as input. Positive rate is calculated for the Chi-square test and optionally for Fisher's exact and the Wald test on an ordinal regression model.
}
\details{
Test
}
\examples{
placeholder
}
