% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Pow_Simul_Mult}
\alias{Pow_Simul_Mult}
\title{Positive Rates for Contingency Tables}
\usage{
Pow_Simul_Mult(
  Simul_Mult_Object = Simul_Mult(),
  add_fisher_exact = FALSE,
  add_ord = FALSE,
  sample_sizes = NULL,
  n_sim = 1000,
  FPR = TRUE
)
}
\arguments{
\item{Simul_Mult_Object}{Output from \code{Simul_Mult()} with the argument \code{verbose} set to TRUE.}

\item{add_fisher_exact}{Whether to compute positive rates for Fisher's Exact test. May add >1 min of calculation time. Defaults to FALSE.}

\item{add_ord}{Whether to compute positive rates for Wald test on a fitted ordinal regression model. May add >1 min of calculation time. Defaults to FALSE.}

\item{sample_sizes}{A vector of sample sizes over which false positive rates are to be calculated. A sample size is defined as the total number of counts in a contingency table. Defaults to total count received by \code{Simul_Mult_Object}.}

\item{n_sim}{Number of contingency tables simulated for each positive rate calculation. Defaults to 1000.}

\item{FPR}{Whether to calculate false positive rate in addition to power. Defaults to TRUE.}
}
\value{
placeholder
}
\description{
Computes the statistical power and, optionally, the false positive rates of tests applied to contingency tables. The contingency table data generation process is defined using \code{Simul_Mult()}, which serves as input. Positive rates are computed for the Chi-square test and optionally for Fisher's exact test and the Wald test applied to an ordinal regression model.
}
\details{
Power is calculated as the percentage of tests that results in positives (p-value < 0.05) when conducted on the simulated set of contingency tables given the data-generating process and parameters like the probability matrix specified in \code{Simul_Mult()}. This probability matrix is supposed to represent the population parameters from which the contingency tables are generated as representatives of our future sample.

The probability [FPR] is set to be equal in each cell for simulations and ultimately calculations of the false positive rate.

Next, talk about each statistical test in the next paragraph.
}
\examples{
placeholder
}
