% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{silencer}
\alias{silencer}
\title{Silence Output}
\usage{
silencer(x)
}
\arguments{
\item{x}{Code which output is to be silenced}
}
\value{
Code output without outputs taken by \code{sink()}, e.g. \code{cat()}.
}
\description{
Hide output from R console by applying \code{sink(tempfile())}
}
\examples{
silencer(a <- cat("test"))

}
