% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Simul_Mult}
\alias{Simul_Mult}
\title{Simulate a Contingency Table}
\usage{
Simul_Mult(
  total_count = 750,
  n_lesion = 3,
  n_Trt. = 5,
  margin_fixed_Trt. = FALSE,
  probs = "equal"
)
}
\arguments{
\item{total_count}{Total number of counts of fish in the contingency table. Default = 750.}

\item{n_lesion}{Number of categories for lesions. Default = 3.}

\item{n_Trt.}{Number of treatment groups. Default = 5.}

\item{margin_fixed_Trt.}{Whether margins are to be fixed per treatment group (i.e. fixed number of fish per treatment). Default = FALSE. See \bold{Details} for further information on marginals.}

\item{probs}{Matrix of probability values created using \code{matrix()}. One row per treatment group and one column per lesion category. All probability values must sum to one if margins_fixed = FALSE (Default). Probability values must sum to one for every treatment group (each row) if margins_fixed = TRUE. Default = equal probability across cells (i.e. for each combination of treatment and lesion category).}
}
\value{
A contingency table with counts for different treatment groups (as rows) and lesion categories (as columns).
}
\description{
Simulate a sample contingency table consisting of counts of fish in \emph{n} lesion category and \emph{n} treatment group. Probability values can be assigned for each combination using \emph{probs}. Function intended for use in assessment of power and/or false positive rates under different experimental conditions.
}
\details{
Counts are simulated based on random sampling from a multinomial distribution (\c?ode{rmultinom()}). Simulation process may assume no fixed marginals or one fixed marginal (per treatment group) in the contingency table.

For explanations and discussion on types of marginals for contingency tables see: \url{https://www.uvm.edu/~statdhtx/StatPages/More_Stuff/Chi-square/Contingency-Tables.pdf} and comments on function arguments.
}
\examples{
Con_Tab = Simul_Mult(total_count = 750, n_lesion = 3, n_Trt. = 5)
}
