% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Simul_Mult}
\alias{Simul_Mult}
\title{Simulate Contingeny Table}
\usage{
Simul_Mult(
  total_count = 750,
  n_lesion = 3,
  n_Trt. = 5,
  margin_fixed_Trt. = FALSE,
  probs = "equal"
)
}
\arguments{
\item{total_count}{Total number of counts of fish in the contingency table. Default = 750.}

\item{n_lesion}{Number of categories for lesions. Default = 3.}

\item{n_Trt.}{Number of treatment groups. Default = 5.}

\item{margin_fixed_Trt.}{Whether margins are to be fixed per treatment group (i.e. fixed number of fish per treatment). Default = FALSE.}

\item{probs}{Matrix of probability values created using 'matrix()'. One row per treatment group and one column per lesion category. All probability values must sum to one if margins_fixed = FALSE (Default). Probability values must sum to one for every treatment group (each row) if margins_fixed = TRUE. Default probability = equal in each cell of the contingency table (i.e. in each combination of treatment group and lesion category).}
}
\description{
Simulates a contingency table consisting of counts of fish in n lesion categories/scores across n treatment groups. Intended for use in assessing power and/or false positive rates under different experimental conditions.
}
\details{
Simulated counts are based on random sampling from a multinomial distribution ('rmultinom()'). Simulation process may assume no fixed marginals or one fixed marginal (per treatment group) in the contingency table. For details on marginals see ref: https://www.uvm.edu/~statdhtx/StatPages/More_Stuff/Chi-square/Contingency-Tables.pdf and comments on function arguments.
}
\examples{
Con_Tab = Simul_Mult(total_count = 750, n_lesion = 3, n_Trt. = 5)
}
