% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Surv_Gen}
\alias{Surv_Gen}
\title{Generate Survivor Data}
\usage{
Surv_Gen(
  mort_db,
  today_tte,
  tank_without_mort = NULL,
  trt_without_mort = NULL,
  starting_fish_count
)
}
\arguments{
\item{mort_db}{A mort data frame described in \bold{Details}.}

\item{today_tte}{The time to event value assigned to the generated rows of survivor data.}

\item{tank_without_mort}{A vector of tank IDs absent from the mort data frame for which survivor data is to be generated.}

\item{trt_without_mort}{A vector of treatment IDs corresponding to \code{tank_without_mort}, in the same order.}

\item{starting_fish_count}{Value representing the starting number of fish in each tank.}
}
\value{
The mort data frame appended with rows representing every survivor.
}
\description{
Produces a completed survival dataset with rows for every survivor based on the starting number of fish and the supplied "mort" data frame. This function can also generate survivor data for tanks not mentioned in the mort data frame (e.g. tanks without mortalities yet) by using the argument \code{tank_without_mort} and \code{trt_without_mort}.
}
\details{
The mort data frame should have one row of data for every dead/sampled fish and the following information in 4 different columns:
\itemize{
\item "Trt.ID" = Label for each treatment group in the study.
\item "Tank.ID" = Label for each different tanks in the study (each tank must have a unique label).
\item "TTE" = Time to Event. The event could be fish death or being sampled and removed depending on "Status".
\item "Status" = Value indicating what happened at TTE. 1 for dead fish 0 for those sampled and removed.
}

For an example dataset, execute \code{data("mort_db_ex")}.
}
