% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Gen}
\alias{Surv_Gen}
\title{Generate Survivor Data}
\usage{
Surv_Gen(
  mort_db,
  starting_fish_count,
  today_tte,
  tank_without_mort = NULL,
  trt_without_mort = NULL
)
}
\arguments{
\item{mort_db}{A mort dataframe as described in \bold{Details}.}

\item{starting_fish_count}{Value representing the starting number of fish per tank.}

\item{today_tte}{Value representing the day or time-to-event the fish survived to. Value assigned to every row of survivor data generated.}

\item{tank_without_mort}{A vector of strings specifying the tanks absent from \code{mort_db} for which to generate survivor data.}

\item{trt_without_mort}{A vector of strings corresponding to \code{tank_without_mort}. Keep the order the same.}
}
\value{
A dataframe produced by combining the input mort data and generated rows of survivor data.
}
\description{
Returns a survival dataframe that includes rows representing every surviving fish based on the starting number of fish provided and the mortality count provided in a dataframe. To generate survivor data for tanks absent in the input dataframe, specify the arguments \code{tank_without_mort} and \code{trt_without_mort}.
}
\details{
The mort dataframe supplied as input should consist of the following 4 columns at minimum:
\itemize{
\item "Trt.ID" = Labels for treatment groups in the study.
\item "Tank.ID" = Labels for tanks in the study (each tank must have a unique label).
\item "TTE" = Time to Event. Event could be fish death or being sampled and removed depending on "Status".
\item "Status" = Value indicating what happened at TTE. 1 for dead fish, 0 for those sampled and removed.
}

Each row should represent one fish. For an example dataset, view \code{data(mort_db_ex)}.
}
\examples{
Surv_Gen(mort_db = mort_db_ex,
         today_tte = 54,
         starting_fish_count = 100,
         tank_without_mort = c("C99", "C100"),
         trt_without_mort = c("A", "B"))

}
