% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Gen}
\alias{Surv_Gen}
\title{Generate Survivor Data}
\usage{
Surv_Gen(
  mort_db,
  starting_fish_count,
  today_tte,
  tank_without_mort = NULL,
  trt_without_mort = NULL
)
}
\arguments{
\item{mort_db}{A mort dataframe as described in \bold{Details}.}

\item{starting_fish_count}{Value representing the starting number of fish for every tank. Alternatively, a dataframe containing the columns "Trt.ID", "Tank.ID", and "starting_fish_count" to allow for different fish starting numbers per tank.}

\item{today_tte}{Value representing the day or time-to-event the fish survived to, assigned to every row of survivor data generated.}

\item{tank_without_mort}{A vector of strings specifying the tanks absent from \code{mort_db}; used to generate survivor data for those tanks. Argument ignored if \code{starting_fish_count} is a dataframe.}

\item{trt_without_mort}{A vector of strings corresponding to \code{tank_without_mort}. Keep their order the same. Argument ignored if \code{starting_fish_count} is a dataframe.}
}
\value{
A dataframe produced by combining the input mort data and generated rows of survivor data.
}
\description{
Produces a survivor dataframe that includes rows representing surviving fish. Number of rows generated depends on the given starting number of fish and mortality data.
}
\details{
The mort dataframe supplied as input should consist of the following 4 columns at minimum:
\itemize{
\item "Trt.ID" = Labels for treatment groups in the study.
\item "Tank.ID" = Labels for tanks in the study (each tank must have a unique label).
\item "TTE" = Time to Event. Event could be fish death or being sampled and removed depending on "Status".
\item "Status" = Value indicating what happened at TTE. 1 for dead fish, 0 for those sampled and removed.
}

Each row should represent one fish.

For an example dataframe, execute \code{data(mort_db_ex)} and view.
}
\examples{
data(mort_db_ex)
Surv_Gen(mort_db = mort_db_ex,
         starting_fish_count = 100,
         today_tte = 54,
         tank_without_mort = c("C99", "C100"),
         trt_without_mort = c("A", "B"))

}
