% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Plots}
\alias{Surv_Plots}
\title{Generate Survival Plots}
\usage{
Surv_Plots(
  surv_db,
  plot_prefix = "plot_prefix",
  x_axis_limits = NULL,
  y_axis_limits = c(0, 1),
  x_lab = "Days Post Challenge",
  lambda1 = NULL
)
}
\arguments{
\item{surv_db}{A survival dataframe as described in \bold{Details}.}

\item{plot_prefix}{A string specifying the prefix for the filename of the saved plots.}

\item{x_axis_limits}{A vector specifying the plot x-axis lower and upper limits, respectively.}

\item{y_axis_limits}{A vector specifying the plot y-axis lower and upper limits, respectively.}

\item{x_lab}{A string specifying the plot x-axis label.}

\item{lambda}{Smoothing parameter for the hazard curve estimated by \code{bshazard()}. Defaults to NULL where \code{bshazard()} employs a data-driven approach to estimating lambda.}
}
\value{
A list containing the Kaplan-Meier Survival Curve and Hazard Curve.
}
\description{
Uses survival data to produce a Kaplan-Meier Survival Plot and a Hazard Plot with curves representing each treatment group. Plots saved automatically to working directory.
}
\details{
The survival dataset should be a dataframe containing at least 4 different columns:
\itemize{
\item "Trt.ID" = Labels for treatment groups in the study.
\item "Tank.ID" = Labels for tanks in the study (each tank must have a unique label).
\item "TTE" = Time to Event. Event depends on "Status".
\item "Status" = Value indicating what happened at TTE. 1 for dead fish, 0 for survivors or those sampled and removed.
}

For an example dataset, view \code{data(surv_db_ex)}.
}
\examples{
Surv_Plots(surv_db = surv_db_ex,
           plot_prefix = "QCATC1090",
           x_axis_limits = c(0, 50),
           y_axis_limits = c(0, 1),
           x_lab = "TTE",
           lambda1 = 10)

}
