% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Plots}
\alias{Surv_Plots}
\title{Generate Survival Plots}
\usage{
Surv_Plots(
  surv_db,
  plot_prefix = "plot_prefix",
  x_axis_limits = NULL,
  y_axis_limits = c(0, 1),
  x_lab = "Days Post Challenge",
  lambda1 = NULL,
  unit = TRUE
)
}
\arguments{
\item{surv_db}{A survival dataframe as described in \bold{Details}.}

\item{plot_prefix}{A string specifying the prefix for the filename of the saved plots.}

\item{x_axis_limits}{A vector specifying the plots x-axis lower and upper limits, respectively.}

\item{y_axis_limits}{A vector specifying the Survival Plot y-axis lower and upper limits, respectively.}

\item{x_lab}{A string specifying the plot x-axis label.}

\item{lambda1}{Smoothing value for the hazard curve estimated by \code{bshazard()}. Defaults to NULL where \code{bshazard()} employs a data-driven approach to estimating lambda. See \bold{Details} if a value is to be chosen.}

\item{unit}{Whether to consider tanks as a fixed covariate in \code{bshazard()} hazard curve estimation. Defaults to TRUE when tank replication is present. When the sample size is high (many morts) and lambda1 is NULL (i.e. lambda is estimated from the data), considering tank as a covariate is highly recommended.}
}
\value{
A list containing the Kaplan-Meier Survival Curve and Hazard Curve.
}
\description{
Uses survival data to produce a Kaplan-Meier Survival Plot and a Hazard Plot with curves representing each treatment group. Plots saved automatically to working directory.
}
\details{
The survival dataset should be a dataframe containing at least 4 different columns:
\itemize{
\item "Trt.ID" = Labels for treatment groups in the study.
\item "Tank.ID" = Labels for tanks in the study (each tank must have a unique label).
\item "TTE" = Time to Event. Event depends on "Status".
\item "Status" = Value indicating what happened at TTE. 1 for dead fish, 0 for survivors or those sampled and removed.
}

For an example dataset, view \code{data(surv_db_ex)}.

For details on the statistical methodology used by \code{bshazard()} to estimate the hazard curve, refer to \url{https://www.researchgate.net/publication /287338889_bshazard_A_Flexible_Tool_for_Nonparametric_Smoothing_of_the_Hazard_Function}. The author first expresses the hazard function h(t) in a count u(t) general linear mixed model; u(t) = h(t) * P(t) where P(t) is the number alive at the given time; the count error/residuals is modeled using a dispersion parameter (phi). The h(t) (basically) is to be modeled using basis splines. By assuming the basis spline curvatures (differences in slopes across hazard values at subsequent time points) are normally distributed with mean 0 (i.e. random effect), the curvatures become a function of the dispersion parameter (phi) estimated from the data. Specifically, variance of curvatures = phi / lambda. The lambda parameter can optionally be specified by the user; if it is increased, the variance of curvature decreases leading to smoother hazard curves, vice versa.

A smoothing parameter of __

When a curve ___ (NULL),

Watch out for ___,
}
\examples{
Surv_Plots(surv_db = surv_db_ex,
           plot_prefix = "QCATC1090",
           x_axis_limits = c(0, 50),
           y_axis_limits = c(0, 1),
           x_lab = "TTE",
           lambda1 = 10)

}
