% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Surv_Plots}
\alias{Surv_Plots}
\title{Title}
\usage{
Surv_Plots(
  surv_db,
  figure_name_prefix = "figure_name_prefix",
  x_axis_limits = NULL,
  y_axis_limits = c(0, 1),
  x_lab = "Days Post Challenge",
  lambda = NULL
)
}
\arguments{
\item{lambda}{}
}
\description{
Title
}
