% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Plots}
\alias{Surv_Plots}
\title{Generate Survival Plots}
\usage{
Surv_Plots(
  surv_db,
  plot_prefix = "plot_prefix",
  xlim = NULL,
  ylim = c(0, 1),
  xlab = "Days Post Challenge",
  lambda = NULL,
  phi = NULL,
  dailybin = TRUE,
  plot = "both"
)
}
\arguments{
\item{surv_db}{A survival dataframe as described in \bold{Details}.}

\item{plot_prefix}{A string specifying the prefix for the filename of the saved plots.}

\item{xlim}{A vector specifying the plots x-axis lower and upper limits, respectively.}

\item{ylim}{A vector specifying the Survival Plot y-axis lower and upper limits, respectively.}

\item{xlab}{A string specifying the plot x-axis label.}

\item{lambda}{Smoothing value for the hazard curve. Higher lambda produces greater smoothing. Defaults to NULL where \code{bshazard()} uses the provided survival data to estimate lambda; recommended for large sample size situations which usually occurs on our full-scale studies with many mortalities and tank-replication. At low sample sizes, the lambda estimate can be unreliable. Choosing a lambda of 10 (or anywhere between 1-100) probably produces the most accurate hazard curve for these situations. In place of choosing lambda, choosing \code{phi} is recommended; see below.}

\item{phi}{Dispersion parameter for the count model used in hazard curve estimation. Defaults to NULL where \code{bshazard()} uses the provided survival data to estimate phi; recommended for large sample size situations. At low sample sizes, the phi estimate can be unreliable. Choosing a phi value of 1 for low sample sizes is recommended. This value of 1 (or close) seems to be that estimated in past Tenaci data (phi ~ 0.8-1.4) in Onda where there are large sample sizes with tank-replication. The value of 1 indicates counts (deaths) vary as "expected" (Poisson) according to the different hazard values along the hazard curve, and are not overdispersed (phi > 1).}

\item{dailybin}{Whether to set time bins at daily (1 TTE) intervals. Refer to the \code{bshazard()} documentation for an understanding on the role of bins to hazard curve estimation. Please set to TRUE at low sample sizes and set to FALSE at large sample sizes with tank-replication. Defaults to TRUE.}

\item{plot}{Which plot to output. Use "surv" for the Kaplan-Meier Survival Curve, "haz" for the Hazard Curve, or "both" for both. Defaults to "both".}
}
\value{
Returns the Kaplan-Meier Survival Curve or Hazard Curve or both in a list depending on the \code{plot} argument.
}
\description{
Uses survival data to produce a Kaplan-Meier Survival Plot and/or Hazard Time Plot. Each plot contains multiple curves for the different treatment groups. Plots saved automatically to working directory.
}
\details{
The survival dataset should be a dataframe containing at least 4 different columns:
\itemize{
\item "Trt.ID" = Labels for treatment groups in the study.
\item "Tank.ID" = Labels for tanks in the study (each tank must have a unique label).
\item "TTE" = Time to Event. Event depends on "Status".
\item "Status" = Value indicating what happened at TTE. 1 for dead fish, 0 for survivors or those sampled and removed.
}

For an example dataset, view \code{data(surv_db_ex)}.

For details on the statistical methodology used by \code{bshazard()}, refer to: \url{https://www.researchgate.net/publication/287338889_bshazard_A_Flexible_Tool_for_Nonparametric_Smoothing_of_the_Hazard_Function}.

General concept: the author first considers h(t) the hazard function in an event-count model: count(t) = h(t) * P(t) where P(t) is the number alive as a function of time. h(t) is modeled over time using basis splines. By assuming the basis spline curvatures have a normal distribution with mean 0 (a random effect), the author found that the curvature's variance can be estimated as a function of the degree of over-dispersion (phi) of counts. The author determined that the variance of curvatures (smoothness) is equal to phi divided by a denominator lambda representing the smoothness parameter. The user can specify phi or lambda which affects the variance of curvatures (smoothness) or can allow them to be estimated from the data.
}
\examples{
Surv_Plots(surv_db = surv_db_ex,
           plot_prefix = "QCATC777",
           xlim = c(0, 50),
           ylim = c(0, 1),
           xlab = "TTE",
           phi = 1,
           plot = "both")

}
