% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Plots}
\alias{Surv_Plots}
\title{Generate Survival Plots}
\usage{
Surv_Plots(
  surv_db,
  plot_prefix = "plot_prefix",
  xlim = NULL,
  ylim = c(0, 1),
  xlab = "Days Post Challenge",
  lambda = NULL,
  phi = NULL,
  dailybin = TRUE,
  plot = "both",
  colours = NULL,
  theme = "ggplot",
  trt_order = NULL,
  data_out = FALSE,
  plot_dim = c(6, 4)
)
}
\arguments{
\item{surv_db}{A survival dataframe as described in \bold{Details}.}

\item{plot_prefix}{A string specifying the prefix for the filename of the saved plots.}

\item{xlim}{A vector specifying the plots x-axis lower and upper limits, respectively.}

\item{ylim}{A vector specifying the Survival Plot y-axis lower and upper limits, respectively.}

\item{xlab}{A string specifying the plot x-axis label.}

\item{lambda}{Smoothing value for the hazard curve. Higher lambda produces greater smoothing. Defaults to NULL where \code{bshazard()} uses the provided survival data to estimate lambda; NULL specification is recommended for large sample size situations which usually occurs on our full-scale studies with many mortalities and tank-replication. At low sample sizes, the lambda estimate can be unreliable. Choosing a lambda of 10 (or anywhere between 1-100) probably produces the most accurate hazard curve for these situations. In place of choosing lambda, choosing \code{phi} is recommended; see below.}

\item{phi}{Dispersion parameter for the count model used in hazard curve estimation. Defaults to NULL where \code{bshazard()} uses the provided survival data to estimate phi; NULL specification is recommended for large sample size situations. At low sample sizes, the phi estimate can be unreliable. Choosing a phi value of 1 for low sample sizes is recommended. This value of 1 (or close) seems to be that estimated in past Tenaci data (QCATC997; phi ~ 0.8-1.4) where there are large sample sizes with tank-replication. The phi value of 1 indicates the set of counts (deaths) over time have a Poisson distribution, following the different hazard rates along the curve and are not overdispersed (phi > 1).}

\item{dailybin}{Whether to set time bins at daily (1 TTE) intervals. Refer to the \code{bshazard()} documentation for an understanding on the role of bins to hazard curve estimation. Please set to TRUE at low sample sizes and set to FALSE for large sample sizes (often with tank replication), although at large sample sizes either TRUE or FALSE produces similar results usually. Defaults to TRUE.}

\item{plot}{Which plot to output. Use "surv" for the Kaplan-Meier Survival Curve, "haz" for the Hazard Curve, or "both" for both. Defaults to "both".}

\item{colours}{Vector of color codes for the different treatment groups in the plot. Defaults to ggplot2 default palette.}

\item{theme}{Character string specifying the graphics theme for the plots. Theme "ggplot2" and "prism" currently available. Defaults to "ggplot2".}

\item{trt_order}{Vector representing the order of treatment groups in the plots. Defaults to NULL where alphabetical order is used.}

\item{data_out}{Whether to print out the survival and/or hazard databases illustrated by the plots. Defaults to FALSE.}

\item{plot_dim}{Vector representing the dimensions (width, height) with which to save the plot in .tiff and .pptx.}
}
\value{
With argument {\code{plot = "both"}}, returns the Kaplan-Meier Survival Curve and the Hazard Curve. Output can be trimmed by setting \code{plot = "haz"} or \code{plot = "surv"}.

If \code{data_out = TRUE}, returns a list of the plot(s) and the associated dataframes to create them.
}
\description{
Produces a Kaplan-Meier Survival Plot and/or Hazard Time Plot from survival data. Each plot contains multiple curves for the different treatment groups. Plots saved automatically to working directory.
}
\details{
The survival dataset should be a dataframe containing at least 4 different columns:
\itemize{
\item "Trt.ID" = Labels for treatment groups in the study.
\item "Tank.ID" = Labels for tanks in the study (each tank must have a unique label).
\item "TTE" = Time to Event. Event depends on "Status".
\item "Status" = Value indicating what happened at TTE. 1 for dead fish, 0 for survivors or those sampled and removed.
}

Each row should represent one fish. For an example dataframe, execute \code{data(surv_db_ex)} and view.

For details on the statistical methodology used by \code{bshazard()}, refer to: \href{https://www.researchgate.net/publication/287338889_bshazard_A_Flexible_Tool_for_Nonparametric_Smoothing_of_the_Hazard_Function}{here}.

General concept: h(t) the hazard function is considered in an count model with the number of deaths as the response variable. I.e, death_count(t) = h(t) * P(t) where P(t) is the number alive as a function of time and h(t) is modeled over time using basis splines. The basis spline curvature\bold{s} is assumed to have a normal distribution with mean 0 (a random effect). Based on this assumption, the author found that the variance of curvatures (i.e. smoothness) is equal to the over-dispersion (phi) of the death counts related (divided) by some smoothness parameter (lambda). Phi and lambda can be estimated from the data or specified by the user. Specification can be helpful in low sample size situations where overdispersion (phi) estimates have been found to be unreliable and clearly wrong (based on my understanding of realistic estimates and what was estimated in past data with adequate, large sample sizes).
}
\examples{
#Starting from an example mortality database, we first generate the complete survivor data using Surv_Gen()
data(mort_db_ex)
surv_dat = Surv_Gen(mort_db = mort_db_ex,
                    starting_fish_count = 100,
                    last_tte = 54)

#Create plot by inputting surv_dat into Surv_Plots()!
Surv_Plots(surv_db = surv_dat,
           plot_prefix = "QCATC777",
           xlim = c(0, 54),
           ylim = c(0, 1),
           xlab = "TTE",
           plot = "both",
           dailybin = FALSE)
}
\seealso{
\href{https://sean4andrew.github.io/safuncs/reference/Surv_Plots.html}{Link} for executed \bold{Examples} which includes any figure outputs.
}
