% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Power}
\alias{Surv_Power}
\title{Calculate Power for Survival Studies}
\usage{
Surv_Power(
  simul_db = simul_db_ex,
  global_test = "logrank",
  model = NULL,
  pairwise_test = "logrank",
  pairwise_corr = "none",
  plot_out = TRUE,
  plot_lines = FALSE,
  xlab = "List Element #",
  xnames = NULL,
  plot_save = TRUE
)
}
\arguments{
\item{simul_db}{The simulated survival dataframe from \code{Surv_Simul()}.}

\item{global_test}{A character vector specifying the method(s) to use for the global hypothesis test of Trt.ID.
\itemize{
\item 'logrank' = the global logrank test. The default for \code{global_test}.
\item 'wald' = the Wald Chisquare test for models. The Wald test can be computed for various models relevant to our experiments and generally produces practically the same p-values as the other two model tests. Hence, the wald test is recommended for model tests. Brief description: the wald test assesses whether model parameters (\emph{log(HR)}) are jointly significantly different from 0 (i.e. \emph{HR} != 1).
\item 'score' = Lagrange multiplier or Score test for models.
\item 'LRT' = Likelihood ratio test for models.
}
For 'wald', 'score', and 'LRT' options, at least one model must be specified in the \code{model} argument.}

\item{model}{A character vector representing the model(s) to fit for hypothesis testing. Models available are: 'coxph_glm' and 'coxph_glmm'. 'coxph_glm' represents the standard cox proportional hazard model fitted using \code{survival::coxph()} with Trt.ID as a fixed factor. 'coxph_glmm' represents the mixed cox proportional hazard model fitted using \code{coxme::coxme()} with Trt.ID as a fixed factor and Tank.ID as a random factor to account for inter-tank variation. Defaults to NULL where no model is used for hypothesis tests.}

\item{pairwise_test}{A character vector representing the method(s) used for pairwise comparisons. Use "logrank" to calculate power for logrank tests comparing different treatments. Use "EMM" to calculate power using Estimated Marginal Means based on model estimates (from 'coxph_glm' and/or 'coxph_glmm'). Defaults to "logrank".}

\item{pairwise_corr}{A character vector representing the method(s) used to adjust p-values for multiplicity of pairwise comparisons. The adjustment method affects power of the pairwise comparisons. Methods available are: "tukey", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "none". In \bold{Details}, I discuss categories of the adjustment methods and provided a recommendation for "BH". Defaults to "none" for now.}

\item{plot_out}{Whether to display plot(s) illustrating power from global and/or pairwise hypothesis tests. Defaults to TRUE}

\item{plot_lines}{Whether to plot lines connecting points of the same "group" in the plot output. Defaults to TRUE.}

\item{xlab}{A string representing the x-axis title. Defaults to "List Element #".}

\item{xnames}{Vector of names for x-axis labels. Defaults to NULL where names are the list element numbers from \code{Surv_Gen()}.}

\item{plot_save}{Whether to save plots as a .tiff. Defaults to TRUE.}

\item{data_out}{Whether to output dataframe(s) containing the power values calculated for the global and/or pairwise hypothesis tests. Defaults to TRUE.}
}
\value{
Output. An SE for proportions (calculated using the binomial formula)
}
\description{
Calculates the power of global and/or pairwise hypothesis tests for survival studies with support over a range of experimental designs. This versatility is enabled by the simulation-based approach of the power calculation, using the modular \code{Surv_Simul()} to simulate survival data from various experimental designs. Power calculations can be made to account for inter-tank variation using a mixed cox proportional hazards model (set the argument \code{model} to "coxph_glmm"). Additionally, power calculations can account for the multiplicity of pairwise comparisons using the \code{pairwise_corr} argument. Users can compare power across different experimental designs by specifying each as a list element in \code{Surv_Simul()}. The results are returned as dataframes and plots.
}
\details{
Power calculation follows the standard procedure for simulation-based approaches. First, the user simulates hypothetical future sample sets using \code{Surv_Simul()}. For each sample set, a p-value is calculated by \code{Surv_Power()}. The percentage of p-values below 0.05 (positives) were then calculated, representing power. The percent positives can also represent false positive rate if the population/truth from which different treatments are simulated are identical.
}
\examples{
Placeholder
}
\seealso{
\href{https://sean4andrew.github.io/safuncs/reference/Surv_Power.html}{Link} for web documentation.
}
