% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Power}
\alias{Surv_Power}
\title{Calculate Power for Survival Experiments}
\usage{
Surv_Power(
  simul_db = simul_db_ex,
  global_test = "logrank",
  model = NULL,
  pairwise_test = "logrank",
  pairwise_corr = "none",
  plot_out = TRUE,
  plot_lines = FALSE,
  xlab = "List Element #",
  xnames = NULL,
  plot_save = TRUE
)
}
\arguments{
\item{simul_db}{The simulated survival dataframe from \code{Surv_Simul()}.}

\item{global_test}{A character vector representing the method(s) used for global hypothesis testing of significance of Trt.ID. Methods available are: 'logrank', 'wald', 'score', 'LRT'. 'logrank' represents the global logrank test of significance. The latter three methods are standard global hypothesis testing methods for model fits. They are only available when the argument \code{model} is specified (i.e. not NULL).'wald' represents the Wald Chisquare Test (also known as joint test) which assesses whether model parameters (log(hazard ratios)) are significantly different than 0 (i.e. HRs =/= 1). This test can be done for various cox-proportional hazards model that could be relevant to our studies (coxph_glm, coxph_glmm, coxph_gee). Due to its wide applicability, while also producing practically the same p-value (most of the time) compared to the other two test for models, 'wald' is recommended of the three. 'score' represents the Lagrange multiplier or Score test. 'LRT' represents the likelihood ratio test. Defaults to 'logrank' for now since it is the most commonly used by our researchers.}

\item{model}{A character vector representing the model(s) to fit for hypothesis testing. Models available are: 'coxph_glm' and 'coxph_glmm'. 'coxph_glm' represents the standard cox proportional hazard model fitted using \code{survival::coxph()} with Trt.ID as a fixed factor. 'coxph_glmm' represents the mixed cox proportional hazard model fitted using \code{coxme::coxme()} with Trt.ID as a fixed factor and Tank.ID as a random factor to account for inter-tank variation. Defaults to NULL where no model is used for hypothesis tests.}

\item{pairwise_test}{A character vector representing the method(s) used for pairwise comparisons. Use "logrank" to calculate power for logrank tests comparing different treatments. Use "EMM" to calculate power using Estimated Marginal Means based on model estimates (from 'coxph_glm' and/or 'coxph_glmm'). Defaults to "logrank".}

\item{pairwise_corr}{A character vector representing the method(s) used to adjust p-values for multiplicity of pairwise comparisons. The adjustment method affects power of the pairwise comparisons. Methods available are: "tukey", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "none". In \bold{Details}, I discuss categories of the adjustment methods and provided a recommendation for "BH". Defaults to "none" for now.}

\item{plot_out}{Whether to display plot(s) illustrating power from global and/or pairwise hypothesis tests. Defaults to TRUE}

\item{plot_lines}{Whether to plot lines connecting points of the same "group" in the plot output. Defaults to TRUE.}

\item{xlab}{A string representing the x-axis title. Defaults to "List Element #".}

\item{xnames}{Vector of names for x-axis labels. Defaults to NULL where names are the list element numbers from \code{Surv_Gen()}.}

\item{plot_save}{Whether to save plots as a .tiff. Defaults to TRUE.}

\item{data_out}{Whether to output dataframe(s) containing the power values calculated for the global and/or pairwise hypothesis tests. Defaults to TRUE.}
}
\value{
Output. An SE for proportions (calculated using the binomial formula)
}
\description{
Calculates power of global and/or pairwise hypothesis tests for survival studies with a wide range of experimental designs. This flexibility is enabled by the simulation-based nature of the power calculation, using the modular \code{Surv_Simul()} to simulate survival data. Power calculations can account for inter-tank variation by using a mixed cox proportional hazards model (set argument \code{model = "coxph_glmm"}). Additionally, power calculations can account for multiplicity of pairwise comparisons using the argument \code{pairwise_corr}. Users may compare power across different experimental designs by specifying each design as list elements in \code{Surv_Simul()}. Power values are returned as dataframes and plots.
}
\details{
The power calculation here follows the standard procedure for power calculation using simulations. Initially, hypothetical future sample sets were simulated using \code{Surv_Simul()}. For each sample set, a p-value is calculated. The percentage of p-values below 0.05 (positives) were then calculated, representing power. The percent positives may also represent false positive rate if the population/truth in the simulation process are identical between treatments.
}
\examples{
Placeholder
}
\seealso{
\href{https://sean4andrew.github.io/safuncs/reference/Surv_Power.html}{Link} for web documentation.
}
