% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Pred}
\alias{Surv_Pred}
\title{Predict Survival Rate}
\usage{
Surv_Pred(
  pred_db,
  ref_db,
  predsr_tte,
  method = 2,
  coxph_mod = "GLMM",
  lambda_pred = NULL,
  phi_pred = NULL
)
}
\arguments{
\item{lambda_pred}{}
}
\description{
Predict survival rate at a specified date given the..
}
\details{
P
}
