% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Pred}
\alias{Surv_Pred}
\title{Predict Future Survival}
\usage{
Surv_Pred(
  surv_db,
  ref_surv_db,
  ref_specs,
  pred_tte,
  dailybin = TRUE,
  phi = 1.5,
  lambda = NULL,
  plot_save = TRUE,
  plot_prefix = "ONDA_XX",
  plot_dim = c(7, 4.3),
  plot_lwrap = FALSE
)
}
\arguments{
\item{surv_db}{A survival dataframe for the ongoing study consisting of at least four columns named TTE, Status, Trt.ID and Tank.ID. For an example, see \code{surv_db_ex}.}

\item{ref_surv_db}{A survival dataframe for the reference group consisting of at least the four column names mentioned in \code{surv_db} documentation. For example, a dataframe loaded form the Survival Data Library.xlsx.}

\item{ref_specs}{A dataframe specifying the study, treatment and time (tte) offset for the reference group in \code{ref_surv_db}. Column names: Study, Trt.ID, TTE_offset. A negative TTE_offset means the reference group survival data times are to be subtracted.}

\item{pred_tte}{A numeric representing the tte at which the survival rate is to be predicted}

\item{dailybin}{Whether to use daily (1 TTE interval) time bins in estimating hazard curves. Defaults to TRUE. Further details in \code{Surv_Plots}.}

\item{phi}{A numeric indicating the count overdispersion parameter to be used in hazard curve estimation. Defaults to 1.5. Further details in \code{Surv_Plots}.}

\item{lambda}{A numeric indicating the smoothing parameter to be used in hazard curve estimation. Defaults to NULL (data driven estimate). Further details in \code{Surv_Plots}.}

\item{plot_save}{Whether to save plot outputs in the working directory. Defaults to TRUE.}

\item{plot_prefix}{A string specifying the prefix of the filename of the saved plots.  Defaults to "ONDA_XX".}

\item{plot_dim}{Numeric vector representing the width and height of the plot (in inches), respectively.}

\item{plot_lwrap}{Whether to apply text wrapping to the "Reference" legend in the plots.}
}
\value{
Outputs a list containing two ggplot2 objects representing the predicted survival curves and predicted hazard curves. Additionally, the list contains a dataframe with the predicted rates over time and another dataframe containing the predicted rates at the specified \code{pred_tte} (or close to it due to limitations in available data from \code{surv_db_ref}). The dataframes have at most 7 columns with the following content description for each: \tabular{lll}{
 \code{ref_unique} \tab \tab The unique reference group used, based on \code{ref_surv_db} and \code{ref_specs}. \cr
 \code{Trt.ID} \tab \tab The treatment group in the ongoing study (\code{surv_db}) with rates to be predicted. \cr
 \code{time} \tab \tab The time to event associated with the hazard rate column (hazard). Round value up (i.e. \code{ceiling()}) to obtain time associated with the survival rate (sr). \cr
 \code{hazard} \tab \tab Hazard rate. \cr
 \code{hr} \tab \tab Hazard ratio (numerator = Trt.ID, denominator = ref_unique). \cr
 \code{cumhaz} \tab \tab Cumulative hazard. \cr
 \code{sr} \tab \tab Survival rate. \cr
}
}
\description{
Predict future survival for an ongoing study (supplied into \code{surv_db}) based on a reference past study (\code{ref_surv_db}). Can consider multiple past studies using the \code{ref_specs} argument. Outputs predicted Kaplan-Meier Survival curves for each treatment group and similarly hazard curves.
}
\details{
Prediction done by firstly estimating a \emph{hazard curve} for the reference group using \code{bshazard::bshazard()}. The ratio of hazards from the past study to the ongoing is estimated using \code{coxme::coxme()}. The reference hazard curve is multiplied by the hazard ratio to obtain the projected hazard curve for the ongoing study. The survival curve is then obtained by exponentiating the negative of the hazard rate.
}
\examples{
# In the first step, we load the reference database then specify the study and
# treatment to use as reference:
ref_surv_db = surv_db_ex
ref_specs = data.frame(Study = "QCATC1068",
                       Trt.ID = "A",
                       TTE_offset = 0)

# Next, we load the ongoing study. Suppose it is mid-way to completion, hence the
# survival data only extends to 35 TTE / DPC and we want to predict to 54. Lets
# create the 35 TTE dataset for this demo:
surv_db = survival::survSplit(data = surv_db_ex[-1,], cut = 35, end = "TTE",
                              event = "Status", episode = "Eps")
surv_db = surv_db[surv_db$Eps == 1, -c(3, 6)]
tail(surv_db, n = 5)

# Now chuck all the created objects into Surv_Pred()!
Surv_Pred(surv_db = surv_db,
          ref_surv_db = ref_surv_db,
          ref_specs = ref_specs,
          pred_tte = 54)
}
\seealso{
\href{https://sean4andrew.github.io/safuncs/reference/Surv_Pred.html}{Link} for executed \bold{Examples} which includes any figure outputs.
}
