% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Pred}
\alias{Surv_Pred}
\title{Predict End Survival Rate}
\usage{
Surv_Pred(
  surv_db,
  ref_surv_db,
  pred_tte,
  pred_method = "simple",
  plot_save = TRUE,
  plot_prefix = "plot_prefix"
)
}
\arguments{
\item{surv_db}{Placeholder}

\item{ref_surv_db}{Placeholder}

\item{pred_tte}{Placeholder}

\item{pred_method}{Placeholder}

\item{plot_save}{Placeholder}

\item{plot_prefix}{Placeholder}
}
\value{
Placeholder
}
\description{
Predict survival rate for a given survival dataset provided a reference survival database used to estimate a reference hazard curve. Prediction done separately by treatment group.
}
\details{
P
}
\examples{
#Placeholder
}
\seealso{
\href{https://sean4andrew.github.io/safuncs/reference/Surv_Pred.html}{Link} for executed \bold{Examples} which includes any figure outputs.
}
