% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Pred}
\alias{Surv_Pred}
\title{Predict Survival Rate}
\usage{
Surv_Pred(
  pred_db,
  ref_db,
  predsr_tte,
  method = 2,
  coxph_mod = "GLMM",
  lambda_pred = NULL,
  phi_pred = NULL
)
}
\arguments{
\item{pred_db}{Placeholder}

\item{ref_db}{Placeholder}

\item{predsr_tte}{Placeholder}

\item{method}{Placeholder}

\item{coxph_mod}{Placeholder}

\item{lambda_pred}{Placeholder}
}
\value{
Placeholder
}
\description{
Predict survival rate for a given survival dataset provided a reference survival database used to estimate a reference hazard curve. Prediction done seperately by treatment group.
}
\details{
P
}
\examples{
Placeholder
}
