% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs2.R
\name{Surv_Pred}
\alias{Surv_Pred}
\title{Predict Survival Rate}
\usage{
Surv_Pred(pred_db, ref_db, predsr_tte, method = 2, coxph_mod = "GLMM")
}
\arguments{
\item{coxph_mod}{}
}
\description{
Predict Survival Rate
}
