% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{Surv_Simul}
\alias{Surv_Simul}
\title{Simulate Survival Data}
\usage{
Surv_Simul(
  haz_db,
  fish_num_per_tank = 100,
  tank_num_per_trt = 4,
  treatments_hr = c(1, 1, 1, 1),
  logHR_sd_intertank = 0.35,
  sample_set = NULL,
  n_simul = 1,
  power = FALSE
)
}
\arguments{
\item{haz_db}{A hazard dataframe representing the reference hazard curve used for simulation of survival data/curves. Can be a dataframe from \code{bshazard()} or \code{Surv_Plots()}. Further information in \bold{Details}.}

\item{fish_num_per_tank}{placeholder}

\item{tank_num_per_trt}{placeholder}

\item{treatments_hr}{placeholder}

\item{logHR_sd_intertank}{placeholder}

\item{sample_set}{placeholder}

\item{n_simul}{placeholder}

\item{power}{placeholder}
}
\value{
placeholder
}
\description{
placeholder
}
\examples{
placeholder
}
