% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safuncs.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Publication theme for ggplot2}
\usage{
theme_Publication(base_size = 14)
}
\arguments{
\item{base_size}{size of text in graph}
}
\value{
theme function to add to a ggplot2 object
}
\description{
A theme function to add to a ggplot2 object for publication style plots. Function adapted from \href{https://rdrr.io/github/HanjoStudy/quotidieR/src/R/theme_publication.R}{HanjoStudy/quotidieR}.
}
\examples{
#Load an example dataset
data(iris)

#Create a ggplot modified with theme_Publication()
library(ggplot2)
ggplot(data = iris, aes(x = Species, colour = Species, y = Petal.Length)) +
   geom_boxplot() +
   theme_Publication()
}
\seealso{
\href{file:///C:/Users/sean4/Documents/GitHub/safuncs/docs/reference/theme_Publication.html}{Link} for executed \bold{Examples} which includes any figure outputs.
}
